change type from decimal to int, so it could be converted from json format when using postman

+++ 04-MVC/YummyRestApi/models/RecipeItem.cs
@@ -12,7 +12,7 @@ namespace YummyRestApi.Models
         public int Time { get; set; }
         public string Difficulty { get; set; }
    
-        public decimal NumberOfLikes { get; set; }
+        public int NumberOfLikes { get; set; }
         public string Ingredients { get; set; }
         public string Process { get; set; }
         public string TipsAndTricks { get; set; }
